+++ src/app/page.tsx
@@ -2,9 +2,5 @@ import Image from "next/image";
 import Logo from "@/assets/tahirahmedt-logo.svg";
 
 export default function Page() {
-  return (
-    <main className="max-w-3xl mx-auto p-8">
-      <Image src={Logo} alt="logo" />
-    </main>
-  );
+  return <main className="max-w-3xl mx-auto p-8">Home page</main>;
 }
