+++ src/components/index.tsx
@@ -1,0 +1,4 @@
+import Header from "./common/Header/Header";
+import Footer from "./common/Footer/Footer";
+
+export { Header, Footer };
